Gentleman introduction to C language
In this part of Fantini's Guide to Programming in C language I saw a lot of things like the basic structure of a C program, streams, ANSI C, GCC , gcc header files, strings and I/O in C language. I think that C language is not hard. Despite that I learned this language a couple years ago...